I have created controllers, models, routes, db and passport
