+++ db/mongoose.js
@@ -1,0 +1,12 @@
+const mongoose = require('mongoose');
+
+mongoose.set('strictQuery', false);
+
+mongoose.connect(process.env.MONGO_SERVER, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+}).then(() => {
+    console.log("Database connected successfully");
+}).catch((error) => {
+    console.log("DB connection failed", process.env.MONGO_SERVER);
+});
