+++ models/User.js
@@ -1,0 +1,20 @@
+const mongoose = require('mongoose');
+const passportLocalMongoose = require('passport-local-mongoose');
+const findOrCreate = require('mongoose-findorcreate');
+
+const userSchema = new mongoose.Schema({
+    username: String,
+    password: String,
+    googleId: String,
+    facebookId: String,
+    secret: [{
+        title: { type: String, required: true },
+        upvote: { type: Number, default: 0 },
+        downvote: { type: Number, default: 0 },
+    }]
+});
+
+userSchema.plugin(passportLocalMongoose);
+userSchema.plugin(findOrCreate);
+
+module.exports = mongoose.model('User', userSchema);
